pyproject.toml
[project]
name = "mcs-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "chainlit>=2.9.0",
    "fastapi>=0.121.2",
    "fastmcp>=2.13.0.2",
    "loguru>=0.7.3",
    "openai>=2.8.0",
    "pydantic-settings>=2.12.0",
    "ruff>=0.14.4",
    "sqlalchemy>=2.0.44",
]
